feat: Add progress bar to OCI upload process
